chore:: move AuthNavigate guard to router folder
