+++ src/pages/LoginPage.tsx
@@ -11,8 +11,8 @@ import { Checkbox } from '@/components/ui/checkbox'
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
+import AuthNavigate from '@/lib/router/AuthNavigate'
 import { LoginSchema } from '@/lib/zod/login-schema'
-import AuthNavigate from '@/pages/AuthNavigate'
 import { useRegisterForm } from '@/util/useRegisterForm'
 
 const Login = () => {
