+++ src/pages/RegisterPage.tsx
@@ -10,8 +10,8 @@ import Progress from '@/components/register/Progress'
 import UserInfoForm from '@/components/register/UserInfoForm'
 import Button from '@/components/ui/button'
 import { Form } from '@/components/ui/form'
+import AuthNavigate from '@/lib/router/AuthNavigate'
 import { RegisterFormSchema } from '@/lib/zod/register-schema'
-import AuthNavigate from '@/pages/AuthNavigate'
 import { ProgressState, RegisterationKey, RegistrationState, ValidState } from '@/types/register'
 import { useRegisterForm } from '@/util/useRegisterForm'
